chore: rename serializer-encapsulation tests, remove smoke-test
