+++ tests/ember-data__serializer/README.md
@@ -1,0 +1,12 @@
+# Tests for @ember-data/serializer
+
+This test-package provides tests for the `@ember-data/serializer` package.
+
+## Running Tests Locally
+
+If you do not already have the project cloned and dependencies installed, you may want to first read [Setting Up The Project](../../contributing/setting-up-the-project.md) 
+
+```cli
+cd tests/ember-data__serializer;
+pnpm test;
+```
