+++ tests/ember-data__serializer/app/index.html
@@ -3,14 +3,14 @@
   <head>
     <meta charset="utf-8" />
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <title>EncapsulationTestApp</title>
+    <title>@ember-data/serializer</title>
     <meta name="description" content="" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
 
     {{content-for "head"}}
 
     <link integrity="" rel="stylesheet" href="{{ rootURL }}assets/vendor.css" />
-    <link integrity="" rel="stylesheet" href="{{ rootURL }}assets/serializer-encapsulation-test-app.css" />
+    <link integrity="" rel="stylesheet" href="{{ rootURL }}assets/ember-data__serializer.css" />
 
     {{content-for "head-footer"}}
   </head>
@@ -18,7 +18,7 @@
     {{content-for "body"}}
 
     <script src="{{ rootURL }}assets/vendor.js"></script>
-    <script src="{{ rootURL }}assets/serializer-encapsulation-test-app.js"></script>
+    <script src="{{ rootURL }}assets/ember-data__serializer.js"></script>
 
     {{content-for "body-footer"}}
   </body>
