+++ tests/ember-data__serializer/tests/index.html
@@ -3,7 +3,7 @@
   <head>
     <meta charset="utf-8" />
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <title>EncapsulationTestApp Tests</title>
+    <title>@ember-data/serializer</title>
     <meta name="description" content="" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
 
@@ -11,7 +11,7 @@
     {{content-for "test-head"}}
 
     <link rel="stylesheet" href="{{ rootURL }}assets/vendor.css" />
-    <link rel="stylesheet" href="{{ rootURL }}assets/serializer-encapsulation-test-app.css" />
+    <link rel="stylesheet" href="{{ rootURL }}assets/ember-data__serializer.css" />
     <link rel="stylesheet" href="{{ rootURL }}assets/test-support.css" />
 
     {{content-for "head-footer"}}
@@ -31,7 +31,7 @@
     <script src="/testem.js" integrity=""></script>
     <script src="{{ rootURL }}assets/vendor.js"></script>
     <script src="{{ rootURL }}assets/test-support.js"></script>
-    <script src="{{ rootURL }}assets/serializer-encapsulation-test-app.js"></script>
+    <script src="{{ rootURL }}assets/ember-data__serializer.js"></script>
     <script src="{{ rootURL }}assets/tests.js"></script>
 
     {{content-for "body-footer"}}
