+++ tests/ember-data__serializer/tests/integration/create-record-test.js
@@ -1,10 +1,11 @@
 import EmberObject from '@ember/object';
 
 import { module, test } from 'qunit';
-import Store from 'serializer-encapsulation-test-app/services/store';
 
+import Store from 'ember-data__serializer/services/store';
 import { setupTest } from 'ember-qunit';
 
+import JSONAPIAdapter from '@ember-data/adapter/json-api';
 import Model, { attr } from '@ember-data/model';
 
 class Person extends Model {
@@ -28,7 +29,7 @@ class Person extends Model {
   }
 }
 
-module('Serializer Contract | serialize methods forward to Serializer#serialize', function (hooks) {
+module('Serializer Contract | running createRecord with minimum serializer', function (hooks) {
   setupTest(hooks);
 
   hooks.beforeEach(function (assert) {
@@ -36,8 +37,9 @@ module('Serializer Contract | serialize methods forward to Serializer#serialize'
     this.owner.register('model:person', Person);
   });
 
-  test('Model#serialize calls Serializer#serialize', async function (assert) {
+  test('save after createRecord calls normalizeResponse and serialize', async function (assert) {
     let serializeCalled = 0;
+    let normalizeResponseCalled = 0;
 
     class TestMinimumSerializer extends EmberObject {
       serialize(snapshot, options) {
@@ -50,16 +52,45 @@ module('Serializer Contract | serialize methods forward to Serializer#serialize'
         const serializedResource = {
           id: snapshot.id,
           type: snapshot.modelName,
-          attributes: {
-            firstName: 'Chris',
-            lastName: 'Thoburn',
-          },
+          attributes: snapshot.attributes(),
         };
 
         return serializedResource;
       }
+
+      normalizeResponse(store, schema, rawPayload, id, requestType) {
+        normalizeResponseCalled++;
+
+        assert.strictEqual(requestType, 'createRecord', 'expected method name is correct');
+        assert.deepEqual(rawPayload, {
+          id: '1m',
+          type: 'person',
+          attributes: {
+            firstName: 'Chris',
+            lastName: 'Thoburn',
+          },
+        });
+
+        return {
+          data: rawPayload,
+        };
+      }
     }
     this.owner.register('serializer:application', TestMinimumSerializer);
+
+    class TestAdapter extends JSONAPIAdapter {
+      ajax(url, type) {
+        return Promise.resolve({
+          id: '1m',
+          type: 'person',
+          attributes: {
+            firstName: 'Chris',
+            lastName: 'Thoburn',
+          },
+        });
+      }
+    }
+    this.owner.register('adapter:application', TestAdapter);
 
     const store = this.owner.lookup('service:store');
 
@@ -78,11 +109,12 @@ module('Serializer Contract | serialize methods forward to Serializer#serialize'
       },
     });
 
-    let serializedPerson = person.serialize();
+    await person.save();
 
+    assert.strictEqual(normalizeResponseCalled, 1, 'normalizeResponse called once');
     assert.strictEqual(serializeCalled, 1, 'serialize called once');
-    assert.deepEqual(serializedPerson, {
-      id: '1',
+    assert.deepEqual(person.toJSON(), {
+      id: '1m',
       type: 'person',
       attributes: {
         firstName: 'Chris',
@@ -91,10 +123,18 @@ module('Serializer Contract | serialize methods forward to Serializer#serialize'
     });
   });
 
-  test('Snapshot#serialize calls Serializer#serialize', async function (assert) {
+  test('save after createRecord calls normalizeResponse and serializeIntoHash if implemented', async function (assert) {
     let serializeCalled = 0;
+    let serializeIntoHashCalled = 0;
+    let normalizeResponseCalled = 0;
 
     class TestMinimumSerializer extends EmberObject {
+      serializeIntoHash(hash, ModelClass, snapshot, options) {
+        serializeIntoHashCalled++;
+
+        hash[snapshot.modelName] = this.serialize(snapshot, options).data;
+      }
+
       serialize(snapshot, options) {
         serializeCalled++;
 
@@ -105,16 +145,49 @@ module('Serializer Contract | serialize methods forward to Serializer#serialize'
         const serializedResource = {
           id: snapshot.id,
           type: snapshot.modelName,
-          attributes: {
-            firstName: 'Chris',
-            lastName: 'Thoburn',
-          },
+          attributes: snapshot.attributes(),
         };
 
         return serializedResource;
       }
+
+      normalizeResponse(store, schema, rawPayload, id, requestType) {
+        normalizeResponseCalled++;
+
+        assert.strictEqual(requestType, 'createRecord', 'expected method name is correct');
+        assert.deepEqual(rawPayload, {
+          person: {
+            id: '1m',
+            type: 'person',
+            attributes: {
+              firstName: 'Chris',
+              lastName: 'Thoburn',
+            },
+          },
+        });
+
+        return {
+          data: rawPayload.person,
+        };
+      }
     }
     this.owner.register('serializer:application', TestMinimumSerializer);
+
+    class TestAdapter extends JSONAPIAdapter {
+      ajax(url, type) {
+        return Promise.resolve({
+          person: {
+            id: '1m',
+            type: 'person',
+            attributes: {
+              firstName: 'Chris',
+              lastName: 'Thoburn',
+            },
+          },
+        });
+      }
+    }
+    this.owner.register('adapter:application', TestAdapter);
 
     const store = this.owner.lookup('service:store');
 
@@ -133,66 +206,13 @@ module('Serializer Contract | serialize methods forward to Serializer#serialize'
       },
     });
 
-    let serializedPerson = person._createSnapshot().serialize();
+    await person.save();
 
+    assert.strictEqual(normalizeResponseCalled, 1, 'normalizeResponse called once');
+    assert.strictEqual(serializeIntoHashCalled, 1, 'serializeIntoHash called once');
     assert.strictEqual(serializeCalled, 1, 'serialize called once');
-    assert.deepEqual(serializedPerson, {
-      id: '1',
-      type: 'person',
-      attributes: {
-        firstName: 'Chris',
-        lastName: 'Thoburn',
-      },
-    });
-  });
-
-  test('Store#serializeRecord calls Serializer#serialize', async function (assert) {
-    let serializeCalled = 0;
-
-    class TestMinimumSerializer extends EmberObject {
-      serialize(snapshot, options) {
-        serializeCalled++;
-
-        assert.strictEqual(snapshot.id, '1', 'id is correct');
-        assert.strictEqual(snapshot.modelName, 'person', 'modelName is correct');
-        assert.deepEqual(snapshot.attributes(), { firstName: 'John', lastName: 'Smith' }, 'attributes are correct');
-
-        const serializedResource = {
-          id: snapshot.id,
-          type: snapshot.modelName,
-          attributes: {
-            firstName: 'Chris',
-            lastName: 'Thoburn',
-          },
-        };
-
-        return serializedResource;
-      }
-    }
-    this.owner.register('serializer:application', TestMinimumSerializer);
-
-    const store = this.owner.lookup('service:store');
-
-    let person = store.createRecord('person', {
-      id: '1',
-      firstName: 'John',
-      lastName: 'Smith',
-    });
-
     assert.deepEqual(person.toJSON(), {
-      id: '1',
-      type: 'person',
-      attributes: {
-        firstName: 'John',
-        lastName: 'Smith',
-      },
-    });
-
-    let serializedPerson = store.serializeRecord(person);
-
-    assert.strictEqual(serializeCalled, 1, 'serialize called once');
-    assert.deepEqual(serializedPerson, {
-      id: '1',
+      id: '1m',
       type: 'person',
       attributes: {
         firstName: 'Chris',
